consolidate calls to ComputeTurbulentViscosity to clean up calling routines
