+++ CMake/BuildERFExe.cmake
@@ -79,6 +79,7 @@ function(build_erf_lib erf_lib_name)
        ${SRC_DIR}/SpatialStencils/StrainRate.H
        ${SRC_DIR}/SpatialStencils/StressTerm.H
        ${SRC_DIR}/SpatialStencils/Interpolation.cpp
+       ${SRC_DIR}/SpatialStencils/ComputeTurbulentViscosity.cpp
        ${SRC_DIR}/SpatialStencils/MomentumToVelocity.cpp
        ${SRC_DIR}/SpatialStencils/VelocityToMomentum.cpp
        ${SRC_DIR}/TimeIntegration/ERF_MRI.H
