+++ Source/ERF_PhysBCFunct.H
@@ -88,23 +88,13 @@ public:
             if (m_has_most_bcs) {
                 const auto& cons_mf = m_data.get_var(Vars::cons);
                 m_viscosity.define(cons_mf.boxArray(), cons_mf.DistributionMap(), EddyDiff::NumDiffs, cons_mf.nGrowVect());
-                AMREX_ALWAYS_ASSERT_WITH_MESSAGE(solverChoice.les_type == LESType::Smagorinsky ||
-                                                 solverChoice.les_type == LESType::Deardorff   ||
-                                                 solverChoice.pbl_type == PBLType::MYNN25,
-                                                 "Must use an LES or PBL model to compute turbulent viscosity for MOST boundaries");
 
-                // PBL or PBL+LES - get vertical viscosity component with PBL model. LES - get it with LES model
-                if (solverChoice.pbl_type != PBLType::None) {
-                    ComputeTurbulentViscosityPBL(m_data.get_var(Vars::xvel),
-                                                 m_data.get_var(Vars::yvel),
-                                                 cons_mf, m_viscosity, geom, solverChoice, m_most);
-                } else {
-                    ComputeTurbulentViscosity(m_data.get_var(Vars::xvel),
-                                              m_data.get_var(Vars::yvel),
-                                              m_data.get_var(Vars::zvel),
-                                              cons_mf, m_viscosity, geom, solverChoice,
-                                              domain_bcs_type_d);
-                }
+                bool vert_only = true;
+                ComputeTurbulentViscosity(m_data.get_var(Vars::xvel),
+                                          m_data.get_var(Vars::yvel),
+                                          m_data.get_var(Vars::zvel),
+                                          cons_mf, m_viscosity, geom, solverChoice, m_most,
+                                          domain_bcs_type_d, vert_only);
             }
         }
